Added support for strings inside single quotes
Now you can escape characters by surrounding them with quotes.
e.g start',||,'end,testing will escape characters inside the quotes
and create a single label start,||,end with action testing

There exists a problem if the first label starts with quotes because it
clashes with the parsing of the parameter in the inner while loop. I
didn't touch that and the solution needs discussing